checking in an experimental feature
new :toggle option.  specify the id of the indicator to toggle the current element with

ex.)
link_to_remote 'link', :url => url, :indicator => {:toggle => indicator_id}
indicator :id => indicator_id

This will toggle the link to the specified indicator when the remote request is made.  ie. The link will be hidden and the indicator will be shown.